Add files via upload
The project is completely based on the idea of developing a user-friendly environment for the users, while being directed towards Thrift item lovers, as the name suggests. Its value is for those who are enthusiasts and are ready to invest in these kinds of products. Some might wonder why this market exists; well, if you do your research, the branded shoe and wardrobe enthusiast market is really huge, and there’s no end to how much they would love such a platform to help them access everything they love all at once. So, I believe yes, this would benefit them, including the Sellers as well, and that is because with this, not only will they be able to reach their own customers in a better environment but also the consumer base of the other stores as well. Similar Ideas can be found on international websites such as Etsy, eBay, etc. These can be expedited into this website by adding new, more recently needed features.